+++ client/src/components/tasks/TasksPlanner.vue
@@ -26,7 +26,7 @@
 
           <ly-list-item
             v-for="i in nbPlaceholders"
-            :key="i"
+            :key="`placeholder-${i}`"
             class="text-secondary"
           >
             {{ $tc(`tasks.planner.importantTaskPlaceholder.${i - 1}`, tasksTotalCount) }}