+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Provide `GET /api/users/me/features` endpoint
 - Accessing `GET /api/users/me` with an invalid user id now returns a `unauthorized` error
 - Provide `DELETE /api/projects/:id` endpoint
+- `PUT /api/tasks/:id/state` endpoint now accepts `started` when task is planned
 
 ## Ara 2018-09-09